requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pip==24.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-telegram-bot==21.10
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0